backend/requirements.txt
# FastAPI Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
aiosqlite>=0.19.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
argon2-cffi>=23.1.0

# Cryptography
cryptography>=42.0.0

# MFA
pyotp>=2.9.0
qrcode[pil]>=7.4.2

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP Client (for HIBP)
httpx>=0.26.0

# Rate Limiting
slowapi>=0.1.9

# Environment
python-dotenv>=1.0.0
